% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_align.R
\name{load_align}
\alias{load_align}
\title{Load .align file}
\usage{
load_align(file)
}
\arguments{
\item{file}{.align file to load}
}
\value{
'load_align()' returns a data frame
}
\description{
In BioNano data, alignment details are saved in .align files. This function
is used to load such data.
}

