% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_alignment.R
\name{load_alignment}
\alias{load_alignment}
\title{Load alignments between optical maps}
\usage{
load_alignment(file, ref_file, qry_file, sort = FALSE,
  expandHitEnum = FALSE)
}
\arguments{
\item{file}{xmap ('xxx.xmap') file to load}

\item{ref_file}{reference cmap ('xxx_r.cmap') file to load}

\item{qry_file}{query cmap ('xxx_q.cmap') file to load}

\item{sort}{logical. If "TRUE", the results will be sorted by
genomic coordinates, otherwise to keep order as in file}

\item{expandHitEnum}{logical. If "TRUE", the 'hitEnum' field will be parsed}
}
\value{
'load_alignment()' returns a list, which contains three data frames,
    named as 'xmap', 'ref' and 'qry'.
}
\description{
BioNano optical maps can be aligned to each other. Normally, the alignment
results are saved into three files, named as "xxx.xmap", "xxx_r.xmap" and
"xxx_q.xmap".
}
\seealso{
load_xmap load_cmap
}

