% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_bed.R
\name{load_bed}
\alias{load_bed}
\title{Load .bed file}
\usage{
load_bed(filename, sort = FALSE)
}
\arguments{
\item{sort}{logical. If "TRUE", the results will be sorted by
             genomic coordinates, otherwise to keep order as in file}

\item{file}{.bed file to load}
}
\value{
'load_bed()' returns a data frame
}
\description{
BioNano structural variants are usually exported as BED files for
visualization in IrysView application.
}
\seealso{
load_smap
}

