% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rbionano.R
\docType{package}
\name{Rbionano-package}
\alias{Rbionano-package}
\title{Rbionano: R package for BioNano data manipulation and visualization}
\description{
Rbionano: R package for BioNano data manipulation and visualization
}
\details{
Rbionano provides functions to manipulate and visualize BioNano data.
}

