% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_align.R
\name{read_align}
\alias{read_align}
\title{Read .align file}
\usage{
read_align(file)
}
\arguments{
\item{file}{.align file to read}
}
\value{
'read_align()' returns a data frame
}
\description{
In BioNano data, alignment details are saved in .align files. This function
is used to read such data.
}

