% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_bed.R
\name{read_bed}
\alias{read_bed}
\title{Read .bed file}
\usage{
read_bed(filename, sort = FALSE)
}
\arguments{
\item{sort}{logical. If "TRUE", the results will be sorted by
             genomic coordinates, otherwise to keep order as in file}

\item{file}{.bed file to read}
}
\value{
'read_bed()' returns a data frame
}
\description{
BioNano structural variants are usually exported as BED files for
visualization in IrysView application.
}
\seealso{
read_smap
}

